move date and lucen utils to utils folder
