+++ src/utils/luceneQueryUtils.js
